Add middleware to switch domains on production
Basically the logic is:

if there is an 'fr' in the path and we're on the english domain,
switch to the french domain.

if there is an 'en' in the path and we're on the french domain,
switch to the english one.

Since the cookies share info now, the user shouldn't notice.