Set .alpha.canada.ca domain for cookies if on Azure
Since our prod service will have 2 URLs pointing at it, we want to
be able to share cookies between the domains so that your session
will be preserved.

Explicitly set the '.alpha.canada.ca' domain in the cookie config
so that user's info can be shared from the french to the english domain
and vice-versa.

ALSO

The way we were setting our cookie config variables were wrong as well
(haha oops), so this PR updates them to the correct syntax.

Sources:
- https://github.com/expressjs/csurf#csurfoptions
- https://github.com/expressjs/cookie-session#options